requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
pyyaml==6.0.1
orjson==3.10.3
rich==13.7.1
tqdm==4.66.4
opencv-python-headless==4.9.0.80
numpy==1.26.4
pandas==2.2.2
prometheus-client==0.20.0
watchdog==4.0.1
cryptography==42.0.7
sqlalchemy==2.0.31
aiohttp==3.9.5
jinja2==3.1.4
requests==2.32.3
